moved isEditMovieOpen and isDeleteMovieOpen inside MovieCard component
