+++ src/components/MovieList/types.tsx
@@ -1,8 +1,7 @@
-import { IHomepageProps } from '@/pages/Homepage/types';
 import { IMoviesListData } from '@/data/MockedDataTypes';
 import { Dispatch, SetStateAction } from 'react';
 
-export interface IMovieListProps extends IHomepageProps {}
+export interface IMovieListProps {}
 
 export interface timesSortedState {
   timesSorted: number;
