+++ src/pages/Homepage/types.ts
@@ -1,6 +1,1 @@
-import { ISetStateBoolean } from 'Global/types/globalTypes';
-
-export interface IHomepageProps {
-  setIsDeleteMovieOpen: ISetStateBoolean;
-  setIsEditMovieOpen: ISetStateBoolean;
-}
+export interface IHomepageProps {}
